requirements.txt
Flask
qrcode
gunicorn
dotenv
gspread
google-auth
